printf

+++ Printf/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: yregragu <yregragu@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/20 17:39:15 by yregragu          #+#    #+#             */
-/*   Updated: 2023/12/21 18:47:34 by yregragu         ###   ########.fr       */
+/*   Updated: 2023/12/22 16:34:17 by yregragu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@
 int	ft_printf(const char *format, ...);
 void	ft_putchar(char c, int *len);
 void   	ft_putstr(char *s, int *len);
-void    ft_putnbr_base(unsigned int n, char c, int *len);
+void    ft_putnbr_base(unsigned long  n, char c, int *len);
 void	ft_putnbr(int n, int *len);
